Wire up clang-tidy with an agreed check list, fix real lint findings
Adds .clang-tidy: bugprone-*/performance-*/portability-*/clang-analyzer-*/
modernize-*/readability-*, with irrelevant house styles (Google/LLVM/
Android/embedded/kernel) and checks verified noisy for this codebase's
own established idioms (implicit-bool-conversion, math-missing-parentheses,
etc.) excluded, each with a written reason checked against real output,
not assumed.

CMakeLists.txt: -Wall -Wextra -Wconversion -Wsign-compare on every target,
-Wno-missing-field-initializers (false-positive-prone for structs with
default member initializers), husk-tests-only -Wno-unused-function
(shared fixture headers reusing names across files with different
signatures -- checked, inline linkage would risk a real ODR violation).

Fixes real findings surfaced by both: 8 narrowing conversions given
explicit static_cast, 6 int*constant widening risks fixed by widening the
loop variable, 9 sign-compare mismatches in tests, 2 unchecked-optional-
access false positives resolved by binding a reference to the optional
before the guard, 2 dead-store initializers (clang-analyzer), 1 empty-
catch and 1 string-concat marked NOLINT with a one-line reason (both
genuinely intentional), plus 74 mechanical use-emplace/use-auto fixes via
clang-tidy --fix, 3 [[nodiscard]] additions, 3 return-braced-init-list
simplifications, 1 raw-string-literal.

modernize-avoid-c-arrays (21 hits, real C-style array struct fields)
deliberately left unfixed -- check stays enabled, flagged as a separate
follow-up rather than bundled into this pass.

Full suite green, 671/671, zero build warnings.

+++ src/dump_chunks_misc.cpp
@@ -260,7 +260,7 @@ void dumpDpiv(json::Writer& w, const Chunk& c) {
     for (size_t i = 0; i < n; ++i) {
         size_t off = i * kSize;
         w.beginObject();
-        for (int f = 0; f < 8; ++f) {
+        for (size_t f = 0; f < 8; ++f) {
             w.key("field_" + std::to_string(f));
             w.value(static_cast<double>(readF32(c.data, c.size, off + f * 4)));
         }
